+++ WebUI/src/app/pipes/battery-state.pipe.ts
@@ -13,8 +13,9 @@ export class BatteryStatePipe implements PipeTransform {
         returnString = "LOW";
     } else if (value == 9999) {
         returnString = "!!!";
-    } else
+    } else {
         returnString = value.toString() + "%";
+    }
 
     return returnString;
   }